:art: Database menu exceeds window https://github.com/siyuan-note/siyuan/issues/17323
